requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
